# docs/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Aiexec
  version: 1.6.4
paths:
...
  /api/v1/mcp/sse:
    get:
      tags:
      - mcp
      summary: Handle Sse
      operationId: handle_sse_api_v1_mcp_sse_get
      responses:
        '200':
          description: Successful Response
      security:
      - OAuth2PasswordBearer: []
      - API key query: []
      - API key header: []
...
  /api/mcp/sse:
    get:
      tags:
      - mcp
      summary: Handle Sse
      operationId: handle_sse_api_mcp_sse_get
      responses:
        '200':
          description: Successful Response
      security:
      - OAuth2PasswordBearer: []
      - API key query: []
      - API key header: []
...
components:
...
  securitySchemes:
    OAuth2PasswordBearer:
      type: oauth2
      flows:
        password:
          scopes: {}
          tokenUrl: api/v1/login
    API key query:
      type: apiKey
      in: query
      name: x-api-key
    API key header:
      type: apiKey
      in: header
      name: x-api-key